Change float expectations to check for numeric instead
* for data values that can be float or integer

+++ spec/facades/weather_facade_spec.rb
@@ -17,8 +17,8 @@ describe WeatherFacade do
     expect(weather.current_weather.datetime).to be_a(String)
     expect(weather.current_weather.sunrise).to be_a(String)
     expect(weather.current_weather.sunset).to be_a(String)
-    expect(weather.current_weather.temperature).to be_a(Float)
-    expect(weather.current_weather.feels_like).to be_a(Float)
+    expect(weather.current_weather.temperature).to be_a(Numeric)
+    expect(weather.current_weather.feels_like).to be_a(Numeric)
     expect(weather.current_weather.conditions).to be_a(String)
     expect(weather.current_weather.icon).to be_a(String)
 
@@ -26,8 +26,8 @@ describe WeatherFacade do
     expect(weather.daily_weather.first.date).to be_a(String)
     expect(weather.daily_weather.first.sunrise).to be_a(String)
     expect(weather.daily_weather.first.sunset).to be_a(String)
-    expect(weather.daily_weather.first.max_temp).to be_a(Float)
-    expect(weather.daily_weather.first.min_temp).to be_a(Float)
+    expect(weather.daily_weather.first.max_temp).to be_a(Numeric)
+    expect(weather.daily_weather.first.min_temp).to be_a(Numeric)
     expect(weather.daily_weather.first.conditions).to be_a(String)
     expect(weather.daily_weather.first.icon).to be_a(String)
